+++ src/parser/enrichers/monster/Generic/_module.mjs
@@ -6,3 +6,4 @@ export { default as PackTactics } from "./PackTactics.mjs";
 export { default as ReversalOfFortune } from "./ReversalOfFortune.mjs";
 export { default as SuaveDefense } from "./SuaveDefense.mjs";
 export { default as NamedEffect } from "./NamedEffect.mjs";
+export { default as Illumination } from "./Illumination.mjs";